requirements.txt
fastapi[standard]
uvicorn
python-dotenv
SQLAlchemy
pydantic
APScheduler
requests
